Strip out `Strategy`, integrate `HasParameterisedGenerator` into `Description`
