+++ examples/Spec/IntCompact.hs
@@ -12,11 +12,11 @@ data IntDescr
   = Zero
   | Positive Size
   | Negative Size
-  deriving stock (Show, Eq, Generic)
+  deriving stock (Show, Eq, Ord, Generic)
   deriving anyclass (SOPGeneric, HasDatatypeInfo)
 
 data Size = Small | Large {isBound :: Bool}
-  deriving stock (Show, Eq, Generic)
+  deriving stock (Show, Eq, Ord, Generic)
   deriving anyclass (SOPGeneric, HasDatatypeInfo)
 
 instance Description IntDescr Int where
@@ -30,23 +30,22 @@ instance Description IntDescr Int where
         | i < 11 && i > -111 = Small
         | otherwise = Large {isBound = i == minBound || i == maxBound}
 
-instance HasParameterisedGenerator (VariableRep IntDescr) Int where
-  parameterisedGenerator Zero = pure 0
-  parameterisedGenerator (Positive (Large True)) = pure maxBound
-  parameterisedGenerator (Positive (Large False)) = int (linear 11 (maxBound - 1))
-  parameterisedGenerator (Positive Small) = int (linear 1 10)
-  parameterisedGenerator (Negative (Large True)) = pure minBound
-  parameterisedGenerator (Negative (Large False)) = int (linear (minBound + 1) (-11))
-  parameterisedGenerator (Negative Small) = int (linear (-10) (-1))
+  descriptionGen Zero = pure 0
+  descriptionGen (Positive (Large True)) = pure maxBound
+  descriptionGen (Positive (Large False)) = int (linear 11 (maxBound - 1))
+  descriptionGen (Positive Small) = int (linear 1 10)
+  descriptionGen (Negative (Large True)) = pure minBound
+  descriptionGen (Negative (Large False)) = int (linear (minBound + 1) (-11))
+  descriptionGen (Negative Small) = int (linear (-10) (-1))
 
 intCompactGenTests :: TestTree
 intCompactGenTests =
   testGroup "intGenTests" $
     fromGroup
-      <$> [ runGeneratorTestsWhere @(VariableRep IntDescr) "Int Generator" Yes
+      <$> [ runGeneratorTestsWhere @IntDescr "Int Generator" Yes
           ]
 
-intCompactPureRunner :: PureRunner (VariableRep IntDescr) Int
+intCompactPureRunner :: PureRunner IntDescr Int
 intCompactPureRunner =
   PureRunner
     { expect = v [("Negative", 0)] "Small"
