+++ examples/Spec/IntSimple.hs
@@ -14,15 +14,15 @@ data IntDescr = IntDescr
   , size :: Size
   , isBound :: Bool
   }
-  deriving stock (Generic, Eq, Show)
+  deriving stock (Generic, Eq, Ord, Show)
   deriving anyclass (SOPGeneric, HasDatatypeInfo)
 
 data Sign = Positive | Negative | Zero
-  deriving stock (Generic, Eq, Show)
+  deriving stock (Generic, Eq, Ord, Show)
   deriving anyclass (SOPGeneric, HasDatatypeInfo)
 
 data Size = Large | Small
-  deriving stock (Generic, Eq, Show)
+  deriving stock (Generic, Eq, Ord, Show)
   deriving anyclass (SOPGeneric, HasDatatypeInfo)
 
 instance Description IntDescr Int where
@@ -46,8 +46,7 @@ instance Description IntDescr Int where
       , v [("IntDescr", "isBound")] "True" :->: v [("IntDescr", "size")] "Large"
       ]
 
-instance HasParameterisedGenerator (VariableRep IntDescr) Int where
-  parameterisedGenerator s =
+  descriptionGen s =
     case sign s of
       Zero -> pure 0
       Positive ->
@@ -67,10 +66,10 @@ intSimpleGenTests :: TestTree
 intSimpleGenTests =
   testGroup "intGenTests" $
     fromGroup
-      <$> [ runGeneratorTestsWhere @(VariableRep IntDescr) "Int Generator" Yes
+      <$> [ runGeneratorTestsWhere @IntDescr "Int Generator" Yes
           ]
 
-intSimplePureRunner :: PureRunner (VariableRep IntDescr) Int
+intSimplePureRunner :: PureRunner IntDescr Int
 intSimplePureRunner =
   PureRunner
     { expect = v [("IntDescr", "size")] "Small" :&&: v [("IntDescr", "sign")] "Negative"
